MDS-197 [ADD] Base Factory Functionality
