+++ modules/app/base/models/base.py
@@ -1,7 +1,7 @@
 from flask import current_app
 from sqlalchemy import or_
 from modules.app.base.controllers.base import BaseController
-from modules.app.base.models.factory import BaseFactory
+from modules.app.base.factories.factory import BaseFactory
 from modules.app.base.routes.base import BaseRoute
 from utils.printable import Printable
 from utils.init_models import init_modules
